CAD: implement deferred driver loading
