+++ frmts/gdalallregister.cpp
@@ -136,6 +136,9 @@ void CPL_STDCALL GDALAllRegister()
 #if defined(DEFERRED_BASISU_KTX2_DRIVER)
     DeclareDeferredBASISU_KTX2Plugin();
 #endif
+#if defined(DEFERRED_CAD_DRIVER)
+    DeclareDeferredOGRCADPlugin();
+#endif
 #if defined(DEFERRED_DDS_DRIVER)
     DeclareDeferredDDSPlugin();
 #endif
